Adds separate Javascript config
Some of the Typescript rules do not work with pure Javascript
projects, especially in Node.js.

So the config is now separated into two files, the Typescript config
extends from the JS config since the rules still partially apply.

Also removed the "root" property as it's about resolution of config
files, not about extending as I thought.

This project itself is still validated using the Typescript config for
now to ensure all configs are valid.